% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRIT.R
\name{cov_RIT}
\alias{cov_RIT}
\title{Coverage-based Random Intersection Trees.}
\usage{
cov_RIT(data, classes, theta, epsilon, epsilon_cat, n_trees = 100L,
  depth = 10L, min_inter_sz = 2L, branch = 5, split_nb = 1, es = TRUE)
}
\arguments{
\item{data}{The dataset as a dataframe.}

\item{classes}{A response vector for the dataset.}

\item{theta}{The prevalence thresholds. Used to determine if an interaction is informative of not w.r.t. the classes.}

\item{epsilon}{The scaling factor for continuous feature removal.}

\item{epsilon_cat}{The scaling factor for categorical feature removal.}

\item{n_trees}{The number of trees.}

\item{depth}{The depth of the interaction search.}

\item{min_inter_sz}{The minimum allowed size of interactions.}

\item{branch}{The branching factor in the interaction search.}

\item{split_nb}{The number of intersections that should be succesively computed before splitting the search into \code{branch} children.}

\item{es}{A logical that indicates whether or not early stopping should be used.}
}
\value{
A model that is composed of all informative interactions found in the dataset.
}
\description{
Executes the algorithm known as Coverage-based Random Intersection Trees.
}
\references{
Ballarini Robin. Random intersection trees for genomic data analysis. Master's thesis, Université Catholique de Louvain, 2016.
}

